Move model arch definitions from factory.py
