+++ src/fairseq2/assets/cards/models/nllb-200.yaml
@@ -5,8 +5,8 @@
 # LICENSE file in the root directory of this source tree.
 
 name: nllb-200
-model_type: nllb  # compat
-model_family: nllb
+model_type: transformer  # compat
+model_family: transformer
 tokenizer: "https://tinyurl.com/flores200sacrebleuspm"
 default_lang: eng_Latn
 langs:
@@ -217,26 +217,26 @@ langs:
 
 name: nllb-200_dense_1b
 base: nllb-200
-model_arch: dense_1b
+model_arch: nllb-200_dense_1b
 checkpoint: "https://tinyurl.com/nllb200dense1bcheckpoint"
 
 ---
 
 name: nllb-200_dense_3b
 base: nllb-200
-model_arch: dense_3b
+model_arch: nllb-200_dense_3b
 checkpoint: "https://tinyurl.com/nllb200dense3bcheckpoint"
 
 ---
 
 name: nllb-200_dense_distill_1b
 base: nllb-200
-model_arch: dense_1b
+model_arch: nllb-200_dense_1b
 checkpoint: "https://tinyurl.com/nllb200densedst1bcheckpoint"
 
 ---
 
 name: nllb-200_dense_distill_600m
 base: nllb-200
-model_arch: dense_600m
+model_arch: nllb-200_dense_600m
 checkpoint: "https://tinyurl.com/nllb200densedst600mcheckpoint"
