+++ src/fairseq2/models/__init__.py
@@ -25,6 +25,7 @@ from fairseq2.models.llama import _register_llama
 from fairseq2.models.mistral import _register_mistral
 from fairseq2.models.nllb import _register_nllb
 from fairseq2.models.s2t_transformer import _register_s2t_transformer
+from fairseq2.models.transformer import _register_transformer
 from fairseq2.models.w2vbert import _register_w2vbert
 from fairseq2.models.wav2vec2 import _register_wav2vec2
 from fairseq2.models.wav2vec2.asr import _register_wav2vec2_asr
@@ -35,6 +36,7 @@ def _register_models() -> None:
     _register_mistral()
     _register_nllb()
     _register_s2t_transformer()
+    _register_transformer()
     _register_w2vbert()
     _register_wav2vec2()
     _register_wav2vec2_asr()
