+++ src/fairseq2/models/s2t_transformer/__init__.py
@@ -4,6 +4,12 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from fairseq2.models.s2t_transformer.archs import (
+    s2t_transformer_arch as s2t_transformer_arch,
+)
+from fairseq2.models.s2t_transformer.archs import (
+    s2t_transformer_archs as s2t_transformer_archs,
+)
 from fairseq2.models.s2t_transformer.factory import (
     S2T_TRANSFORMER_FAMILY as S2T_TRANSFORMER_FAMILY,
 )
@@ -16,25 +22,19 @@ from fairseq2.models.s2t_transformer.factory import (
 from fairseq2.models.s2t_transformer.factory import (
     create_s2t_transformer_model as create_s2t_transformer_model,
 )
-from fairseq2.models.s2t_transformer.factory import (
-    s2t_transformer_arch as s2t_transformer_arch,
-)
-from fairseq2.models.s2t_transformer.factory import (
-    s2t_transformer_archs as s2t_transformer_archs,
-)
 from fairseq2.models.s2t_transformer.feature_extractor import (
     Conv1dFbankSubsampler as Conv1dFbankSubsampler,
 )
 from fairseq2.models.s2t_transformer.frontend import (
     S2TTransformerFrontend as S2TTransformerFrontend,
 )
-from fairseq2.models.s2t_transformer.setup import (
+from fairseq2.models.s2t_transformer.loader import (
     load_s2t_transformer_config as load_s2t_transformer_config,
 )
-from fairseq2.models.s2t_transformer.setup import (
+from fairseq2.models.s2t_transformer.loader import (
     load_s2t_transformer_model as load_s2t_transformer_model,
 )
-from fairseq2.models.s2t_transformer.setup import (
+from fairseq2.models.s2t_transformer.loader import (
     load_s2t_transformer_tokenizer as load_s2t_transformer_tokenizer,
 )
 from fairseq2.models.s2t_transformer.tokenizer import (
@@ -45,9 +45,12 @@ from fairseq2.models.s2t_transformer.tokenizer import (
 
 from fairseq2.data.text import load_text_tokenizer
 from fairseq2.models.loader import load_model
+from fairseq2.models.s2t_transformer.archs import _register_s2t_transformer_archs
 
 
 def _register_s2t_transformer() -> None:
+    _register_s2t_transformer_archs()
+
     load_model.register(S2T_TRANSFORMER_FAMILY, load_s2t_transformer_model)
 
     load_text_tokenizer.register(S2T_TRANSFORMER_FAMILY, load_s2t_transformer_tokenizer)
