+++ src/fairseq2/models/wav2vec2/__init__.py
@@ -4,6 +4,14 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from fairseq2.models.wav2vec2.archs import wav2vec2_arch as wav2vec2_arch
+from fairseq2.models.wav2vec2.archs import wav2vec2_archs as wav2vec2_archs
+from fairseq2.models.wav2vec2.archs import (
+    wav2vec2_encoder_arch as wav2vec2_encoder_arch,
+)
+from fairseq2.models.wav2vec2.archs import (
+    wav2vec2_encoder_archs as wav2vec2_encoder_archs,
+)
 from fairseq2.models.wav2vec2.factory import WAV2VEC2_FAMILY as WAV2VEC2_FAMILY
 from fairseq2.models.wav2vec2.factory import Wav2Vec2Builder as Wav2Vec2Builder
 from fairseq2.models.wav2vec2.factory import Wav2Vec2Config as Wav2Vec2Config
@@ -16,14 +24,6 @@ from fairseq2.models.wav2vec2.factory import (
 from fairseq2.models.wav2vec2.factory import (
     create_wav2vec2_model as create_wav2vec2_model,
 )
-from fairseq2.models.wav2vec2.factory import wav2vec2_arch as wav2vec2_arch
-from fairseq2.models.wav2vec2.factory import wav2vec2_archs as wav2vec2_archs
-from fairseq2.models.wav2vec2.factory import (
-    wav2vec2_encoder_arch as wav2vec2_encoder_arch,
-)
-from fairseq2.models.wav2vec2.factory import (
-    wav2vec2_encoder_archs as wav2vec2_encoder_archs,
-)
 from fairseq2.models.wav2vec2.feature_extractor import (
     Wav2Vec2FbankFeatureExtractor as Wav2Vec2FbankFeatureExtractor,
 )
@@ -31,6 +31,8 @@ from fairseq2.models.wav2vec2.feature_extractor import (
     Wav2Vec2FeatureExtractor as Wav2Vec2FeatureExtractor,
 )
 from fairseq2.models.wav2vec2.frontend import Wav2Vec2Frontend as Wav2Vec2Frontend
+from fairseq2.models.wav2vec2.loader import load_wav2vec2_config as load_wav2vec2_config
+from fairseq2.models.wav2vec2.loader import load_wav2vec2_model as load_wav2vec2_model
 from fairseq2.models.wav2vec2.masker import Wav2Vec2Masker as Wav2Vec2Masker
 from fairseq2.models.wav2vec2.model import Wav2Vec2Features as Wav2Vec2Features
 from fairseq2.models.wav2vec2.model import Wav2Vec2Loss as Wav2Vec2Loss
@@ -42,13 +44,14 @@ from fairseq2.models.wav2vec2.position_encoder import (
 from fairseq2.models.wav2vec2.position_encoder import (
     Wav2Vec2StackedPositionEncoder as Wav2Vec2StackedPositionEncoder,
 )
-from fairseq2.models.wav2vec2.setup import load_wav2vec2_config as load_wav2vec2_config
-from fairseq2.models.wav2vec2.setup import load_wav2vec2_model as load_wav2vec2_model
 
 # isort: split
 
 from fairseq2.models.loader import load_model
+from fairseq2.models.wav2vec2.archs import _register_wav2vec2_archs
 
 
 def _register_wav2vec2() -> None:
+    _register_wav2vec2_archs()
+
     load_model.register(WAV2VEC2_FAMILY, load_wav2vec2_model)
