+++ src/fairseq2/models/wav2vec2/asr/__init__.py
@@ -4,6 +4,8 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+from fairseq2.models.wav2vec2.asr.archs import wav2vec2_asr_arch as wav2vec2_asr_arch
+from fairseq2.models.wav2vec2.asr.archs import wav2vec2_asr_archs as wav2vec2_asr_archs
 from fairseq2.models.wav2vec2.asr.factory import (
     WAV2VEC2_ASR_FAMILY as WAV2VEC2_ASR_FAMILY,
 )
@@ -14,23 +16,22 @@ from fairseq2.models.wav2vec2.asr.factory import Wav2Vec2AsrConfig as Wav2Vec2As
 from fairseq2.models.wav2vec2.asr.factory import (
     create_wav2vec2_asr_model as create_wav2vec2_asr_model,
 )
-from fairseq2.models.wav2vec2.asr.factory import wav2vec2_asr_arch as wav2vec2_asr_arch
-from fairseq2.models.wav2vec2.asr.factory import (
-    wav2vec2_asr_archs as wav2vec2_asr_archs,
+from fairseq2.models.wav2vec2.asr.loader import (
+    load_wav2vec2_asr_config as load_wav2vec2_asr_config,
+)
+from fairseq2.models.wav2vec2.asr.loader import (
+    load_wav2vec2_asr_model as load_wav2vec2_asr_model,
 )
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrModel as Wav2Vec2AsrModel
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrOutput as Wav2Vec2AsrOutput
-from fairseq2.models.wav2vec2.asr.setup import (
-    load_wav2vec2_asr_config as load_wav2vec2_asr_config,
-)
-from fairseq2.models.wav2vec2.asr.setup import (
-    load_wav2vec2_asr_model as load_wav2vec2_asr_model,
-)
 
 # isort: split
 
 from fairseq2.models.loader import load_model
+from fairseq2.models.wav2vec2.asr.archs import _register_wav2vec2_asr_archs
 
 
 def _register_wav2vec2_asr() -> None:
+    _register_wav2vec2_asr_archs()
+
     load_model.register(WAV2VEC2_ASR_FAMILY, load_wav2vec2_asr_model)
