+++ src/fairseq2/models/wav2vec2/asr/factory.py
@@ -7,52 +7,15 @@
 from dataclasses import dataclass, field
 from typing import Final, Optional
 
-from fairseq2.config_registry import ConfigRegistry
 from fairseq2.models.wav2vec2.asr.model import Wav2Vec2AsrModel
 from fairseq2.models.wav2vec2.factory import (
     Wav2Vec2EncoderBuilder,
     Wav2Vec2EncoderConfig,
-    wav2vec2_encoder_archs,
 )
 from fairseq2.models.wav2vec2.masker import Wav2Vec2Masker
 from fairseq2.typing import DataType, Device
 
 WAV2VEC2_ASR_FAMILY: Final = "wav2vec2_asr"
-
-
-def _base_encoder() -> Wav2Vec2EncoderConfig:
-    config = wav2vec2_encoder_archs.get("base")
-
-    config.feature_gradient_scale = 1.0
-    config.dropout_p = 0.0
-    config.attn_dropout_p = 0.0
-    config.ffn_inner_dropout_p = 0.1
-
-    return config
-
-
-def _large_encoder() -> Wav2Vec2EncoderConfig:
-    config = wav2vec2_encoder_archs.get("large")
-
-    config.feature_gradient_scale = 1.0
-    config.dropout_p = 0.0
-    config.attn_dropout_p = 0.0
-    config.ffn_inner_dropout_p = 0.1
-    config.layer_drop_p = 0.1
-
-    return config
-
-
-def _large_lv60k_encoder() -> Wav2Vec2EncoderConfig:  # LibriVox 60K
-    config = wav2vec2_encoder_archs.get("large_lv60k")
-
-    config.feature_gradient_scale = 1.0
-    config.dropout_p = 0.0
-    config.attn_dropout_p = 0.0
-    config.ffn_inner_dropout_p = 0.1
-    config.layer_drop_p = 0.1
-
-    return config
 
 
 @dataclass
@@ -63,7 +26,14 @@ class Wav2Vec2AsrConfig:
     :cite:t:`https://doi.org/10.48550/arxiv.2006.11477`.
     """
 
-    encoder_config: Wav2Vec2EncoderConfig = field(default_factory=_base_encoder)
+    encoder_config: Wav2Vec2EncoderConfig = field(
+        default_factory=lambda: Wav2Vec2EncoderConfig(
+            feature_gradient_scale=1.0,
+            dropout_p=0.0,
+            attn_dropout_p=0.0,
+            ffn_inner_dropout_p=0.1,
+        )
+    )
     """The configuration of the encoder."""
 
     final_dim: int = 32
@@ -95,69 +65,6 @@ class Wav2Vec2AsrConfig:
 
     min_num_spatial_mask_spans: int = 2
     """The minimum number of spatial masks sampled per sequence."""
-
-
-wav2vec2_asr_archs = ConfigRegistry[Wav2Vec2AsrConfig]()
-
-wav2vec2_asr_arch = wav2vec2_asr_archs.decorator
-
-
-@wav2vec2_asr_arch("base_10h")
-def _base_10h() -> Wav2Vec2AsrConfig:
-    return Wav2Vec2AsrConfig()
-
-
-@wav2vec2_asr_arch("base_100h")
-def _base_100h() -> Wav2Vec2AsrConfig:
-    config = _base_10h()
-
-    config.encoder_config.layer_drop_p = 0.1
-
-    return config
-
-
-@wav2vec2_asr_arch("large_10h")
-def _large_10h() -> Wav2Vec2AsrConfig:
-    config = _base_10h()
-
-    config.encoder_config = _large_encoder()
-    config.max_temporal_mask_prob = 0.80
-    config.max_spatial_mask_prob = 0.30
-
-    return config
-
-
-@wav2vec2_asr_arch("large_100h")
-def _large_100h() -> Wav2Vec2AsrConfig:
-    config = _base_10h()
-
-    config.encoder_config = _large_encoder()
-    config.max_temporal_mask_prob = 0.53
-    config.max_spatial_mask_prob = 0.55
-
-    return config
-
-
-@wav2vec2_asr_arch("large_lv60k_10h")  # LibriVox 60K
-def _large_lv60k_10h() -> Wav2Vec2AsrConfig:
-    config = _base_10h()
-
-    config.encoder_config = _large_lv60k_encoder()
-    config.max_temporal_mask_prob = 0.80
-    config.max_spatial_mask_prob = 0.30
-
-    return config
-
-
-@wav2vec2_asr_arch("large_lv60k_100h")  # LibriVox 60K
-def _large_lv60k_100h() -> Wav2Vec2AsrConfig:
-    config = _base_10h()
-
-    config.encoder_config = _large_lv60k_encoder()
-    config.max_temporal_mask_prob = 0.53
-    config.max_spatial_mask_prob = 0.55
-
-    return config
 
 
 class Wav2Vec2AsrBuilder:
